fixes to vulnerability elements UI

+++ mapcomposer/app/static/externals/gxp/src/script/plugins/SyntheticView.js
@@ -88,6 +88,7 @@ gxp.plugins.SyntheticView = Ext.extend(gxp.plugins.Tool, {
     notVisibleOnGridMessage: "Formula non visibile a questa scala",
     simMsg: 'Modifica dei parametri di simulazione non possibile a questa scala. Zoomare fino a scala 1:17061',
     downloadFileLabel: 'Scarica il file',
+    deleteDownloadError: 'Il download non può essere cancellato. Rimuoverlo ugualmente?',
     // End i18n.
         
     id: "syntheticview",
@@ -525,6 +526,40 @@ gxp.plugins.SyntheticView = Ext.extend(gxp.plugins.Tool, {
             }
         }
         return null;
+    },
+    
+    deleteDownload: function(downloadUrl, removeResource, callback) {
+        var me = this;
+        var deleteDownloadProcess = this.wpsClient.getProcess('destination', 'gs:DestinationRemoveDownload');                                  
+                                
+        deleteDownloadProcess.execute({
+            headers: me.geoStoreUser ? {
+                "Authorization":  "Basic " + Base64.encode(me.geoStoreUser + ":" + me.geoStorePassword)
+            } : undefined,
+            // spatial input can be a feature or a geometry or an array of
+            // features or geometries
+            inputs: {
+                url: new OpenLayers.WPSProcess.LiteralData({value:downloadUrl})
+            },
+            outputs: [],                                    
+            success: function(outputs) {
+                
+                if(outputs.executeResponse.status.processSucceeded) {
+                    var success = outputs.executeResponse.processOutputs[0].literalData.value === 'true';
+                    
+                    if(callback) {
+                        callback.call(me, success, me.deleteDownloadError);
+                    }
+                } else {
+                    var error = outputs.executeResponse.status.exception.exceptionReport.exceptions[0].texts[0]
+                    
+                    if(callback) {
+                        callback.call(me, false, error);
+                    }                    
+                }
+            }
+        });
+        
     },
     
     /** private: method[addOutput]
@@ -1335,7 +1370,7 @@ gxp.plugins.SyntheticView = Ext.extend(gxp.plugins.Tool, {
                                 } else {
                                     bounds = map.getExtent();
                                 }         
-                                
+                                var scale = me.getScaleFromBounds(bounds);
                                 var filter = new OpenLayers.Filter.Spatial({ 
                                   type: OpenLayers.Filter.Spatial.BBOX,
                                   property: 'geometria',
@@ -1387,6 +1422,7 @@ gxp.plugins.SyntheticView = Ext.extend(gxp.plugins.Tool, {
                                         distanceNames: new OpenLayers.WPSProcess.LiteralData({value: distanceNames.join(',')}),
                                         fp: new OpenLayers.WPSProcess.LiteralData({value:status.temporal.value}),
                                         language: new OpenLayers.WPSProcess.LiteralData({value:GeoExt.Lang.locale}),
+                                        onlyarcs: new OpenLayers.WPSProcess.LiteralData({value:scale > me.analiticViewScale}),
                                         damageArea: status.processing === 4 ? new OpenLayers.WPSProcess.LiteralData({value:status.damageArea}) : undefined,
                                         cff: cff,
                                         padr: padr,
@@ -1692,10 +1728,31 @@ gxp.plugins.SyntheticView = Ext.extend(gxp.plugins.Tool, {
                                                                             type: 'resource',
                                                                             id: id
                                                                         };
-                                                                        store.remove(record);
+                                                                        
                                                                         gpanel.getSelectionModel().clearSelections(true);
+                                                                        var downloadUrl = me.newDownloadStatus;
                                                                         me.newDownloadStatus = null;
-                                                                        me.geoStore.deleteEntity(removeResource);
+                                                                        me.deleteDownload.call(me, downloadUrl, removeResource, function(success, errorMsg) {
+                                                                            if(success) {
+                                                                                 me.geoStore.deleteEntity(removeResource);
+                                                                                 store.remove(record);
+                                                                            } else {
+                                                                                Ext.Msg.show({
+                                                                                    title: me.removeProcessingMsgTitle,
+                                                                                    buttons: Ext.Msg.YESNO,
+                                                                                    msg: errorMsg,
+                                                                                    fn: function(btn, text) {
+                                                                                        if(btn == 'yes') {
+                                                                                            me.geoStore.deleteEntity(removeResource);
+                                                                                            store.remove(record);
+                                                                                        }
+                                                                                    },
+                                                                                    icon: Ext.MessageBox.WARNING,
+                                                                                    scope: this
+                                                                                });
+                                                                            }
+                                                                        });
+                                                                        
                                                                     }
                                                                 }
                                                                 
@@ -2512,22 +2569,22 @@ gxp.plugins.SyntheticView = Ext.extend(gxp.plugins.Tool, {
         }, true));
     },*/
 
-    addVulnLayer: function(layers,layer,formulaDesc, group) {
-        this.currentRiskLayers.push(layer);
-        var group = group;
-        var viewParams = "bounds:264455.272103\\,4889918.130753\\,594775.925407\\,5120964.085926;residenti:1;turistica:1;industria:1;sanitarie:1;scolastiche:1;commerciali:1;urbanizzate:0;boscate:0;protette:0;agricole:0;sotterranee:0;superficiali:0;culturali:0;sostanze:1\\,2\\,3\\,4\\,5\\,6\\,7\\,8\\,9\\,10;scenari:1\\,2\\,3\\,4\\,5\\,6\\,7\\,8\\,9\\,10\\,11;gravita:0\\,1";
-        var env = "low:100;medium:500;max:1000;formula:32;target:98;materials:1,2,3,4,5,6,7,8,9,10;scenarios:1,2,3,4,5,6,7,8,9,10,11;entities:0,1;fp:fp_scen_centrale;processing:1";
-        layers.push(this.createLayerRecord({
-            name: layer,
-            title: formulaDesc, 
+    addVulnLayer: function(layers,layerName,title, group) {
+        if(this.vulnerabilityLayer) {
+            this.removeLayersByName(this.target.mapPanel.map,[this.vulnerabilityLayer]);
+        }
+        this.vulnerabilityLayer = layerName;
+        var env = "coverages:"+layers.join(',');
+        var record = this.createLayerRecord({
+            name: layerName,
+            title: title, 
             tiled: false,
             params: {                                                                
-                viewparams: viewParams,
-                env: env,
-                defaultenv: env,
-                riskPanel: true
+                env: env
             }
-        }, true, undefined, undefined, group));
+        }, true, undefined, undefined, group);
+        
+        this.target.mapPanel.layers.add([record]);
     },
     
     addFormula: function(layers, bounds, status, targetId, layer, formulaDesc, formulaUdm, env) {
@@ -3083,6 +3140,13 @@ gxp.plugins.SyntheticView = Ext.extend(gxp.plugins.Tool, {
     
     getMapScale: function() {
         return Math.round(this.target.mapPanel.map.getScale());        
+    },
+    
+    getScaleFromBounds: function(bounds) {
+        var map = this.target.mapPanel.map;
+        var zoom = map.getZoomForExtent(bounds, true);
+        var res = map.getResolutionForZoom(zoom);
+        return OpenLayers.Util.getScaleFromResolution(res, map.baseLayer.units);
     },
     
     getChosenTarget: function(status) {
